prov/verbs: Separate MR functionality from Domain
