+++ prov/verbs/Makefile.include
@@ -8,6 +8,7 @@ _verbs_files =							\
 	prov/verbs/src/verbs_cq.c				\
 	prov/verbs/src/verbs_srq.c				\
 	prov/verbs/src/verbs_domain.c				\
+	prov/verbs/src/verbs_mr.c				\
 	prov/verbs/src/verbs_eq.c				\
 	prov/verbs/src/verbs_info.c				\
 	prov/verbs/src/verbs_msg.c				\
